Move NumberDataType
