+++ misc/python/materialize/output_consistency/input_data/types/number_types_provider.py
@@ -7,7 +7,30 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
 
-from materialize.output_consistency.data_type.data_type import NumberDataType
+from typing import Optional
+
+from materialize.output_consistency.data_type.data_type import DataType
+from materialize.output_consistency.data_type.data_type_category import DataTypeCategory
+
+
+class NumberDataType(DataType):
+    def __init__(
+        self,
+        identifier: str,
+        type_name: str,
+        is_signed: bool,
+        is_decimal: bool,
+        tiny_value: str,
+        max_value: str,
+        max_negative_value: Optional[str],
+    ):
+        super().__init__(identifier, type_name, DataTypeCategory.NUMERIC)
+        self.is_signed = is_signed
+        self.is_decimal = is_decimal
+        self.tiny_value = tiny_value
+        self.max_value = max_value
+        self.max_negative_value = max_negative_value
+
 
 INT2_TYPE = NumberDataType(
     "INT2",
